+++ src/Kerbalism/Integrations/CryoTanks/CryoTankAccess.cs
@@ -12,5 +12,7 @@ namespace KERBALISM
 		internal static string GetFuelName(object fuelEntry) => CryoTanks.GetFuelName(fuelEntry);
 
 		internal static float GetBoiloffRate(object fuelEntry) => CryoTanks.GetBoiloffRate(fuelEntry);
+
+		internal static float GetFuelCoolingCost(object fuelEntry) => CryoTanks.GetFuelCoolingCost(fuelEntry);
 	}
 }